+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/popup/newControl/NewControlPopup.java
@@ -11,11 +11,9 @@
 package com.kaylerrenslow.armaDialogCreator.gui.fx.main.popup.newControl;
 
 import com.kaylerrenslow.armaDialogCreator.arma.control.impl.ArmaControlLookup;
-import com.kaylerrenslow.armaDialogCreator.control.ControlClass;
-import com.kaylerrenslow.armaDialogCreator.control.ControlProperty;
-import com.kaylerrenslow.armaDialogCreator.control.ControlPropertyLookup;
-import com.kaylerrenslow.armaDialogCreator.control.ControlType;
+import com.kaylerrenslow.armaDialogCreator.control.*;
 import com.kaylerrenslow.armaDialogCreator.control.sv.SerializableValue;
+import com.kaylerrenslow.armaDialogCreator.data.ApplicationDataManager;
 import com.kaylerrenslow.armaDialogCreator.data.io.export.ProjectExporter;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.BorderedImageView;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.inputfield.IdentifierChecker;
@@ -143,6 +141,13 @@ public class NewControlPopup extends StagePopup<VBox> {
 		}
 
 		updatePreview();
+	}
+
+	@Override
+	protected void ok() {
+		ApplicationDataManager.getInstance().getCurrentProject().getCustomControlClassRegistry().addControlClass(new ControlClassSpecification(editorPane.getControlClass()));
+
+		super.ok();
 	}
 
 	@Override