+++ CHANGES.md
@@ -1,5 +1,5 @@
 **Added:**
-* control class specification-this will be for saving and loading custom control classes
+* xml loading and writing changes
 
 **Changed:**
 * none